refactor(P-048): extract DateField into shared component; use in finance form
Moves the custom calendar picker and its helpers out of TariffEditor into
src/components/DateField.tsx so it can be reused. FinanceForm now uses
the same DateField component for investment date instead of a native input.
